pyproject.toml
[tool.poetry]
name = "straw"
version = "0.1.0"
description = ""
authors = ["Gautham Kamath"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {version = "0.154.1", extras = ["debug-server"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
